requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
huggingface-hub>=0.20.1
Pillow==10.2.0
python-dotenv==1.0.1
pydantic==2.5.3
requests==2.31.0
typing-extensions==4.8.0